Break patches and transforms out of Composition types
We have a lot of logic defined as methods on our API types; this breaks
them out into distinct files to make it a little clearer which is which.
