+++ apis/apiextensions/v1/composition_transforms_test.go
@@ -1,0 +1,269 @@
+/*
+Copyright 2020 The Crossplane Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package v1
+
+import (
+	"reflect"
+	"testing"
+
+	"github.com/google/go-cmp/cmp"
+	"github.com/pkg/errors"
+
+	"github.com/crossplane/crossplane-runtime/pkg/test"
+)
+
+func TestMapResolve(t *testing.T) {
+	type args struct {
+		m map[string]string
+		i interface{}
+	}
+	type want struct {
+		o   interface{}
+		err error
+	}
+
+	cases := map[string]struct {
+		args
+		want
+	}{
+		"NonStringInput": {
+			args: args{
+				i: 5,
+			},
+			want: want{
+				err: errors.Errorf(errFmtMapTypeNotSupported, "int"),
+			},
+		},
+		"KeyNotFound": {
+			args: args{
+				i: "ola",
+			},
+			want: want{
+				err: errors.Errorf(errFmtMapNotFound, "ola"),
+			},
+		},
+		"Success": {
+			args: args{
+				m: map[string]string{"ola": "voila"},
+				i: "ola",
+			},
+			want: want{
+				o: "voila",
+			},
+		},
+	}
+	for name, tc := range cases {
+		t.Run(name, func(t *testing.T) {
+			got, err := (&MapTransform{Pairs: tc.m}).Resolve(tc.i)
+
+			if diff := cmp.Diff(tc.want.o, got); diff != "" {
+				t.Errorf("Resolve(b): -want, +got:\n%s", diff)
+			}
+			if diff := cmp.Diff(tc.want.err, err, test.EquateErrors()); diff != "" {
+				t.Errorf("Resolve(b): -want, +got:\n%s", diff)
+			}
+		})
+	}
+}
+
+func TestMathResolve(t *testing.T) {
+	m := int64(2)
+
+	type args struct {
+		multiplier *int64
+		i          interface{}
+	}
+	type want struct {
+		o   interface{}
+		err error
+	}
+
+	cases := map[string]struct {
+		args
+		want
+	}{
+		"NoMultiplier": {
+			args: args{
+				i: 25,
+			},
+			want: want{
+				err: errors.New(errMathNoMultiplier),
+			},
+		},
+		"NonNumberInput": {
+			args: args{
+				multiplier: &m,
+				i:          "ola",
+			},
+			want: want{
+				err: errors.New(errMathInputNonNumber),
+			},
+		},
+		"Success": {
+			args: args{
+				multiplier: &m,
+				i:          3,
+			},
+			want: want{
+				o: 3 * m,
+			},
+		},
+		"SuccessInt64": {
+			args: args{
+				multiplier: &m,
+				i:          int64(3),
+			},
+			want: want{
+				o: 3 * m,
+			},
+		},
+	}
+	for name, tc := range cases {
+		t.Run(name, func(t *testing.T) {
+			got, err := (&MathTransform{Multiply: tc.multiplier}).Resolve(tc.i)
+
+			if diff := cmp.Diff(tc.want.o, got); diff != "" {
+				t.Errorf("Resolve(b): -want, +got:\n%s", diff)
+			}
+			if diff := cmp.Diff(tc.want.err, err, test.EquateErrors()); diff != "" {
+				t.Errorf("Resolve(b): -want, +got:\n%s", diff)
+			}
+		})
+	}
+}
+
+func TestStringResolve(t *testing.T) {
+
+	type args struct {
+		fmts string
+		i    interface{}
+	}
+	type want struct {
+		o   interface{}
+		err error
+	}
+
+	cases := map[string]struct {
+		args
+		want
+	}{
+		"FmtString": {
+			args: args{
+				fmts: "verycool%s",
+				i:    "thing",
+			},
+			want: want{
+				o: "verycoolthing",
+			},
+		},
+		"FmtInteger": {
+			args: args{
+				fmts: "the largest %d",
+				i:    8,
+			},
+			want: want{
+				o: "the largest 8",
+			},
+		},
+	}
+	for name, tc := range cases {
+		t.Run(name, func(t *testing.T) {
+			got, err := (&StringTransform{Format: tc.fmts}).Resolve(tc.i)
+
+			if diff := cmp.Diff(tc.want.o, got); diff != "" {
+				t.Errorf("Resolve(b): -want, +got:\n%s", diff)
+			}
+			if diff := cmp.Diff(tc.want.err, err, test.EquateErrors()); diff != "" {
+				t.Errorf("Resolve(b): -want, +got:\n%s", diff)
+			}
+		})
+	}
+}
+
+func TestConvertResolve(t *testing.T) {
+	type args struct {
+		ot string
+		i  interface{}
+	}
+	type want struct {
+		o   interface{}
+		err error
+	}
+
+	cases := map[string]struct {
+		args
+		want
+	}{
+		"StringToBool": {
+			args: args{
+				i:  "true",
+				ot: ConvertTransformTypeBool,
+			},
+			want: want{
+				o: true,
+			},
+		},
+		"SameTypeNoOp": {
+			args: args{
+				i:  true,
+				ot: ConvertTransformTypeBool,
+			},
+			want: want{
+				o: true,
+			},
+		},
+		"IntAliasToInt64": {
+			args: args{
+				i:  int64(1),
+				ot: ConvertTransformTypeInt,
+			},
+			want: want{
+				o: int64(1),
+			},
+		},
+		"InputTypeNotSupported": {
+			args: args{
+				i:  []int{64},
+				ot: ConvertTransformTypeString,
+			},
+			want: want{
+				err: errors.Errorf(errFmtConvertInputTypeNotSupported, reflect.TypeOf([]int{}).Kind().String()),
+			},
+		},
+		"ConversionPairNotSupported": {
+			args: args{
+				i:  "[64]",
+				ot: "[]int",
+			},
+			want: want{
+				err: errors.Errorf(errFmtConversionPairNotSupported, "string", "[]int"),
+			},
+		},
+	}
+	for name, tc := range cases {
+		t.Run(name, func(t *testing.T) {
+			got, err := (&ConvertTransform{ToType: tc.args.ot}).Resolve(tc.i)
+
+			if diff := cmp.Diff(tc.want.o, got); diff != "" {
+				t.Errorf("Resolve(b): -want, +got:\n%s", diff)
+			}
+			if diff := cmp.Diff(tc.want.err, err, test.EquateErrors()); diff != "" {
+				t.Errorf("Resolve(b): -want, +got:\n%s", diff)
+			}
+		})
+	}
+}
